requirements.txt
attrs==21.4.0
click==8.0.3
colorama==0.4.4
decorator==5.1.1
flasgger==0.9.5
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
flask-msearch==0.2.9.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mistune==2.0.2
PyJWT==2.3.0
pyrsistent==0.18.1
python-dotenv==0.19.2
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.31
validators==0.18.2
Werkzeug==2.0.2
zipp==3.7.0
